sui_kvstore

Trait KeyValueStoreReader

pub trait KeyValueStoreReader {
    // Required methods
    fn get_objects<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        objects: &'life1 [ObjectKey],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_transactions<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        transactions: &'life1 [TransactionDigest],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_checkpoints<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        sequence_numbers: &'life1 [CheckpointSequenceNumber],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<CheckpointData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_checkpoint_by_digest<'life0, 'async_trait>(
        &'life0 mut self,
        digest: CheckpointDigest,
    ) -> Pin<Box<dyn Future<Output = Result<Option<CheckpointData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_watermark_for_pipelines<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        pipelines: &'life1 [&'life2 str],
    ) -> Pin<Box<dyn Future<Output = Result<Option<Watermark>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn get_latest_object<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        object_id: &'life1 ObjectID,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_epoch<'life0, 'async_trait>(
        &'life0 mut self,
        epoch_id: EpochId,
    ) -> Pin<Box<dyn Future<Output = Result<Option<EpochData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_latest_epoch<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<EpochData>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_events_for_transactions<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        keys: &'life1 [TransactionDigest],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<(TransactionDigest, TransactionEventsData)>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_object_types<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        object_ids: &'life1 [ObjectID],
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ObjectType>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;

    // Provided method
    fn get_watermark<'life0, 'async_trait>(
        &'life0 mut self,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Watermark>>> + Send + 'async_trait>>
       where Self: Send + 'async_trait,
             'life0: 'async_trait { ... }
}

Required Methods

fn get_objects<'life0, 'life1, 'async_trait>( &'life0 mut self, objects: &'life1 [ObjectKey], ) -> Pin<Box<dyn Future<Output = Result<Vec<Object>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_transactions<'life0, 'life1, 'async_trait>( &'life0 mut self, transactions: &'life1 [TransactionDigest], ) -> Pin<Box<dyn Future<Output = Result<Vec<TransactionData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_checkpoints<'life0, 'life1, 'async_trait>( &'life0 mut self, sequence_numbers: &'life1 [CheckpointSequenceNumber], ) -> Pin<Box<dyn Future<Output = Result<Vec<CheckpointData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_checkpoint_by_digest<'life0, 'async_trait>( &'life0 mut self, digest: CheckpointDigest, ) -> Pin<Box<dyn Future<Output = Result<Option<CheckpointData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_watermark_for_pipelines<'life0, 'life1, 'life2, 'async_trait>( &'life0 mut self, pipelines: &'life1 [&'life2 str], ) -> Pin<Box<dyn Future<Output = Result<Option<Watermark>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Return the minimum watermark across the given pipelines, selecting the whole watermark with the lowest checkpoint_hi_inclusive. Returns None if any pipeline is missing a watermark.

fn get_latest_object<'life0, 'life1, 'async_trait>( &'life0 mut self, object_id: &'life1 ObjectID, ) -> Pin<Box<dyn Future<Output = Result<Option<Object>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_epoch<'life0, 'async_trait>( &'life0 mut self, epoch_id: EpochId, ) -> Pin<Box<dyn Future<Output = Result<Option<EpochData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_latest_epoch<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<Option<EpochData>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_events_for_transactions<'life0, 'life1, 'async_trait>( &'life0 mut self, keys: &'life1 [TransactionDigest], ) -> Pin<Box<dyn Future<Output = Result<Vec<(TransactionDigest, TransactionEventsData)>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn get_object_types<'life0, 'life1, 'async_trait>( &'life0 mut self, object_ids: &'life1 [ObjectID], ) -> Pin<Box<dyn Future<Output = Result<Vec<ObjectType>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Provided Methods

fn get_watermark<'life0, 'async_trait>( &'life0 mut self, ) -> Pin<Box<dyn Future<Output = Result<Option<Watermark>>> + Send + 'async_trait>>
where Self: Send + 'async_trait, 'life0: 'async_trait,

Return the minimum watermark across all non-legacy pipelines.